rename globalTypes to globals
