+++ src/index.ts
@@ -1,4 +1,4 @@
-import './globalTypes.js'; // makes global types available to user code
+import './globals.js'; // makes global types available to user code
 
 export type {Task} from './task/task.js';
 export {TaskError} from './task/task.js';
