+++ tsconfig.json
@@ -23,7 +23,7 @@
     // "composite": true, // Enable project compilation: https://www.typescriptlang.org/docs/handbook/project-references.html
     // "diagnostics": false, // Show diagnostic information.
     // "incremental": true, // Enable incremental compilation by reading/writing information from prior compilations to a file on disk.
-    // TODO turn `isolatedModules` on once TS4.3 lands and `exports` are added and `globalTypes` is redesigned
+    // TODO turn `isolatedModules` on once TS4.3 lands and `exports` are added and `globals` is redesigned
     "isolatedModules": false, // Transpile each file as a separate module (similar to 'ts.transpileModule').
     "importsNotUsedAsValues": "remove", // This flag controls how imports work. When set to `remove`, imports that only reference types are dropped. When set to `preserve`, imports are never dropped. When set to `error`, imports that can be replaced with `import type` will result in a compiler error. Requires TypeScript version 3.8 or later.
     // "listEmittedFiles": false, // Print names of generated files part of the compilation.
